pyproject.toml
[tool.poetry]
name = "ts-media-bridge"
version = "0.1.0"
description = "Bridge Spotify and NYT APIs to analyze streaming popularity vs media coverage."
authors = ["Sofia Giorgianni <sg3925@columbia.edu>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "ts_media_bridge" }]

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.31.0"
pandas = "^2.0.0"
python-dotenv = "^1.0.0"
jupyterlab = ">=4.5.0,<5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
sphinx = "^8.0.0"
sphinx-rtd-theme = "^3.0.0"
sphinx-autodoc-typehints = "^2.0.0"

[tool.poetry.extras]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "sphinx-autodoc-typehints",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

